6350 | 18 files | M DumpSource2/schemas/client/CDOTA_Ability_Kez_Echo_Slash.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Dawnbreaker_Celestial_Hammer_Movement.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Snapfire_Magma_Thinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Snapfire_SpitCreep_ArcingUnit.h,  M DumpSource2/schemas/client/CDOTA_Modifier_TemplarAssassin_Trap.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Visage_SummonFamiliars_StoneForm_Buff.h,  M DumpSource2/schemas/client/CDOTA_Modifier_VoidSpirit_AetherRemnant_Unit.h,  M DumpSource2/schemas/server/CDOTA_Ability_Kez_Echo_Slash.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Dawnbreaker_Celestial_Hammer_Movement.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Snapfire_Magma_Thinker.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Snapfire_SpitCreep_ArcingUnit.h,  M DumpSource2/schemas/server/CDOTA_Modifier_TemplarAssassin_Trap.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Visage_SummonFamiliars_StoneForm_Buff.h,  M DumpSource2/schemas/server/CDOTA_Modifie
https://steamdb.info/patchnotes/17669168/

+++ DumpSource2/schemas/client/CDOTA_Modifier_Snapfire_Magma_Thinker.h
@@ -1,7 +1,7 @@
 class CDOTA_Modifier_Snapfire_Magma_Thinker
 {
 	float32 burn_interval;
-	int32 impact_radius;
+	float32 impact_radius;
 	int32 move_slow_pct;
 	float32 burn_linger_duration;
 	ParticleIndex_t m_nPathEffectIndex;
