+++ DumpSource2/schemas/client/CDOTA_Modifier_TemplarAssassin_Trap.h
@@ -1,4 +1,4 @@
 class CDOTA_Modifier_TemplarAssassin_Trap
 {
-	int32 bonus_vision;
+	float32 bonus_vision;
 };
