+++ DumpSource2/schemas/server/CDOTA_Modifier_Snapfire_SpitCreep_ArcingUnit.h
@@ -13,10 +13,10 @@ class CDOTA_Modifier_Snapfire_SpitCreep_ArcingUnit
 	float32 m_fClampedProjectileSpeed;
 	ParticleIndex_t m_nFXIndex;
 	float32 m_fAcceleration;
-	int32 min_range;
+	float32 min_range;
 	float32 min_lob_travel_time;
 	float32 max_lob_travel_time;
-	int32 impact_radius;
+	float32 impact_radius;
 	int32 projectile_vision;
 	float32 stun_duration;
 	float32 min_height_above_lowest;
