+++ DumpSource2/schemas/server/CDOTA_Modifier_TemplarAssassin_Trap.h
@@ -1,5 +1,5 @@
 class CDOTA_Modifier_TemplarAssassin_Trap
 {
 	CUtlVector< CHandle< CBaseEntity > > m_vecTriggers;
-	int32 bonus_vision;
+	float32 bonus_vision;
 };
