+++ DumpSource2/schemas/server/CDOTA_Modifier_Visage_SummonFamiliars_StoneForm_Buff.h
@@ -1,5 +1,5 @@
 class CDOTA_Modifier_Visage_SummonFamiliars_StoneForm_Buff
 {
 	float32 hp_regen;
-	int32 stun_radius;
+	float32 stun_radius;
 };
